Get rid of hdf5_storage.
We removed reliance on hdf5_storage in SeisCL.py because it was redundant with h5py and was problematic with the latest version of h5py. We had to make sure that dimension are in the same ordering: SeisCL uses column major arrays wheres hdf5 format is row major. This requires transposing the numpy arrays before writing them to disk.